apps/nsqd: re-organize
